# Improve performance of scanning source files